improve error handling for keepalive connections
Problem: When the client crashes during a transaction, the fsm is stuck in an already started transaction.
So when the client tries to start a new transaction, the fsm crashes and the connection is no longer usable if it is KeepAlive=true.

Fix: When the clocksi_interactivetx_coord_fsm crashes, it will send an error to
the process that started the transaction.
If the connection is using KeepAlive it will try to restart the transaction with a new fsm.