Added monitoring interface for Prometheus
This allows to monitor Antidote by polling on
<antidotehost>:3001/metrics using a Prometheus server.
The port is hard-coded at the moment.

The metrics collected at the moment are basically the Erlang-VM metrics.
A connected Grafana server can show these metrics using the
BEAM-Dashboard available here:
https://github.com/deadtrickster/beam-dashboards/blob/master/BEAM-dashboard.json

The staleness metric is not yet converted, but this should be easy
enough.