Prometheus Monitoring (#321)
* Added monitoring interface for Prometheus

This allows to monitor Antidote by polling on
<antidotehost>:3001/metrics using a Prometheus server.
The port is hard-coded at the moment.

The metrics collected at the moment are basically the Erlang-VM metrics.
A connected Grafana server can show these metrics using the
BEAM-Dashboard available here:
https://github.com/deadtrickster/beam-dashboards/blob/master/BEAM-dashboard.json

The staleness metric is not yet converted, but this should be easy
enough.

* ported exisiting staleness property

staleness is not exported as Prometheus metrics. Additional metrics
are work in progress (e.g. open transactions, aborted transactions)

* tools for setting up monitoring

* made metrics port a config parameter

* more metrics and updated Dashboard

* error monitor

* dashboard update

* better units in Dashboard

* expose antidote port

* use 0 as default limit for Dashboard

* Stack operation in Dashboard