Adds an API to read or validate an object.
This adds an function to the antidote module that allows a client
to cache some read resut.

The API takes a arbitrary token as an input, if the token is valid
the client can use the cached value, otherwise the new value is returned
alongside a token for the next call.

* A token is the commit time of a snapshot, if commit time is different
  a new value is returned. An empty token is a niche value that is
  always invalid.

* When the commit time for a snapshot is unknown, the token is considered
  invalid.

* The materializer has been modified a bit to return the CommitTime alongside
  the snapshot.

* The materializer is the one who does the check to avoid unecessary data
  roundtrips.

* Transactions with local updates always invalidate the token if updates
  concerns the same partition. This use case is considered rare so no
  optimisation have been tried on this.